+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   added hover-over breakdown for weapon attack ratings
 -   added ability to disable split damage results
+-   added ability to filter damage types
 -   added link to create an issue in footer
 
 ### Changed